test/mochitest/test_device-eng_stop_b2g.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1233029
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container">
  <h3 id="offline"></h3>
  <ul id="devices"></ul>
  <input id="userBuild" type="file" value="Upload your own build" />
  <button id="installBtn">Flash Device</button>
  <div class="dialog" id="confirmDialog"><div>
    <section>
      <p>Your device has succesfully installed Firefox OS!</p>
       <button>Ok</button>
    </section>
  </div></div>
</div>
<pre id="test">
  <script type="application/javascript" id="code"></script>
  <script type="application/javascript" id="fake-adb"></script>
  <script type="application/javascript">

  "use strict";

  var fakeAdbDevice;
  var realAlert = window.alert;
  var mockAlertCalled = false;
  var mockAlert = function(msg) {
    console.debug(msg);
    mockAlertCalled = true;
    return;
  };

  function populateFakeDevices() {
    return loadFakeDevice().then(() => {
      fakeAdbDevice = new FakeAdbDevice();
      fakeAdbDevice._props["ro.build.type"] = "eng";
      fakeAdbDevice._props["ro.debuggable"] = "1";
      fakeAdbDevice._services["b2g"] = true;
    });
  }

  function test_stopB2G() {
    return new Promise((resolve, reject) => {
      is(fakeAdbDevice._services["b2g"], true, "b2g service running");
      stopB2GOrDisableLockScreen(fakeAdbDevice).then(() => {
        is(mockAlertCalled, false, "window.alert() not called");
        is(fakeAdbDevice._services["b2g"], false, "b2g service stopped");
        ok(true, "stopB2GOrDisableLockScreen() succeeded");
        return resolve();
      }).catch(() => {
        ok(false, "stopB2GOrDisableLockScreen() failed");
        return reject();
      });
    });
  }

  function runTest() {
    loadAboutJs()
      .then(() => populateFakeDevices())
      .then(() => {
        window.alert = mockAlert;
      })
      .then(() => test_stopB2G())
      .then(() => {
        window.alert = realAlert;
      })
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
